[BE] MOVE: PersistHelper 패키지 변경
